+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.6] - 2019-08-10
 ### Added
 - Limitation of maximum project drafts.
 - Project view counting.
@@ -41,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First deployment version of this project.
 - This CHANGELOG file even this is not a open source project.
 
-[Unreleased]: https://github.com/ccmikechen/Birdiy-server/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/ccmikechen/Birdiy-server/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.6
 [0.0.5]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.5
 [0.0.4]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.4
 [0.0.3]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.3